+++ mitra_utils/src/urls.rs
@@ -1,10 +1,7 @@
 use std::net::{Ipv4Addr, Ipv6Addr};
 
-use percent_encoding::{
-    percent_decode_str,
-    utf8_percent_encode,
-    NON_ALPHANUMERIC,
-};
+use iri_string::percent_encode::PercentEncodedForUri;
+use percent_encoding::percent_decode_str;
 use url::{Host, Url};
 
 pub use url::{
@@ -16,8 +13,10 @@ pub fn validate_uri(uri: &str) -> Result<(), UrlError> {
     Ok(())
 }
 
+/// Encode URI path component (RFC-3986).
+/// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/encodeURIComponent#encoding_for_rfc3986
 pub fn url_encode(input: &str) -> String {
-    utf8_percent_encode(input, NON_ALPHANUMERIC).to_string()
+    PercentEncodedForUri::unreserve(input).to_string()
 }
 
 pub fn url_decode(input: &str) -> String {
@@ -124,9 +123,33 @@ mod tests {
 
     #[test]
     fn test_url_encode_decode() {
-        let input = "El Niño";
+        let input = "El Niño"; // unicode and space
         let output = url_encode(input);
         assert_eq!(output, "El%20Ni%C3%B1o");
+        let decoded = url_decode(&output);
+        assert_eq!(decoded, input);
+    }
+
+    #[test]
+    fn test_url_encode_decode_reserved_characters() {
+        let input_1 = ";/?:@&=+$,#"; // encoded by encodeURIComponent()
+        let encoded_1 = url_encode(input_1);
+        assert_eq!(encoded_1, "%3B%2F%3F%3A%40%26%3D%2B%24%2C%23");
+        let decoded_1 = url_decode(&encoded_1);
+        assert_eq!(decoded_1, input_1);
+
+        let input_2 = "-.!~*'()"; // not encoded by encodeURIComponent()
+        let encoded_2 = url_encode(input_2);
+        assert_eq!(encoded_2, "-.%21~%2A%27%28%29");
+        let decoded_2 = url_decode(&encoded_2);
+        assert_eq!(decoded_2, input_2);
+    }
+
+    #[test]
+    fn test_url_encode_decode_url() {
+        let input = "https://social.example/users/test_user";
+        let output = url_encode(input);
+        assert_eq!(output, "https%3A%2F%2Fsocial.example%2Fusers%2Ftest_user");
         let decoded = url_decode(&output);
         assert_eq!(decoded, input);
     }